The main change of this version is the async and sync serial operation. In the pass time,we only can use serial sync OR async but not AND. That means when we read the receive bytes at RxHandler,we should not read the bytes at the other thread except for the RxHandler.If not ,we got some receive parts at RxHandler,and got some bytes at other thread's Read function. If this situation happened,we hard to parse the receive bytes.
When we can not except when the data will come in,use RxHandler is better.
When we CAN EXCEPT some data is coming,use Read(timeout) out of RxHandler will
be better.

More case is,some times we kown that something is coming,and we want to Read()
at the same thread,and some times we waitting for some unexcept datas.
How can we get the balance of sync and async?

That is the highlight of the commit,set the difference READ_MODE,you can got
the effect you need.

Use SetReadMode(RMODE_ASYNC),the object will work like the other serial class,
RxHandler and other thread can Read() at the same time,my class do nothing.

Use RMODE_SYNC,we only can Read() by Read() at the other thread,RxHandler will
not be called,of cause,this is not ASYNC mode.

Use RMODE_AUTO,like RMODE_ASYNC,we can Read() outside and inside RxHandler.
The wonderful is when we Read() outside,the RxHandler will not rob the receive
bytes we wanted.And if we don't Read() outside,RxHandler will remind us.